+++ evm_arithmetization/tests/add11_yml.rs
@@ -232,41 +232,3 @@ fn add11_yml() -> anyhow::Result<()> {
 
     verify_all_proofs(&all_stark, &proofs, &config)
 }
-
-/// This test focuses on testing zkVM proofs with some empty tables.
-#[test]
-fn add11_yml_first_segment() -> anyhow::Result<()> {
-    type F = GoldilocksField;
-    const D: usize = 2;
-    type C = PoseidonGoldilocksConfig;
-
-    init_logger();
-
-    let all_stark = AllStark::<F, D>::default();
-    let config = StarkConfig::standard_fast_config();
-    let inputs = get_generation_inputs();
-
-    let mut timing = TimingTree::new("prove", log::Level::Debug);
-
-    let max_cpu_len_log = 7;
-    let mut segment_iterator = SegmentDataIterator::<F>::new(&inputs, Some(max_cpu_len_log));
-    let mut proofs = vec![];
-    let (_, mut segment_data) = segment_iterator.next().unwrap()?;
-
-    let opcode_counts = &segment_data.opcode_counts;
-    assert!(!opcode_counts.contains_key(&KeccakGeneral));
-
-    let proof = prove::<F, C, D>(
-        &all_stark,
-        &config,
-        inputs.trim(),
-        &mut segment_data,
-        &mut timing,
-        None,
-    )?;
-    proofs.push(proof);
-
-    timing.filter(Duration::from_millis(100)).print();
-
-    verify_all_proofs(&all_stark, &proofs, &config)
-}
